Use requirements.txt to pin nightly versions.
Such that we are not at mercy of the status of the most recent nightly.

Also, switch to py3.8 on GPU machines, since the TPU 1VM images use
py3.8, and we cannot override that.

The current nightly is pinned to 20220718, which is near or at the TPU 1VM image's
nightly version. The date is found from `pip list |grep tf-`, which
revealed in the TPU 1VM image the tf-estimator-nightly is built on this date.

We shall update this once the TPU 1VM image is updated. The CloudTPU
team is actively working to bring the 1vm image up to date.